Change Key Value type to uint64_t

+++ Prototype/include/common.h
@@ -1,7 +1,11 @@
 #include <cstdint>
 #pragma once
-typedef int32_t Key_t; // 8 bytes
-typedef int32_t Val_t; // 8 bytes
+//typedef int32_t Key_t; // 8 bytes
+//typedef int32_t Val_t; // 8 bytes
+
+typedef uint64_t Key_t; // 8 bytes
+typedef uint64_t Val_t; // 8 bytes
+
 const int MAX_LEVEL = 16;
 #define SEARCH_STABLITY_COEFFICIENT 4
 #define WORKERQUEUE_NUM 1
